+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.245] - 2022-02-23
+- [Print stacktrace link in REPL gets duplicated](https://github.com/BetterThanTomorrow/calva/issues/1542)
+- [Publish pre-releases when dev updates](https://github.com/BetterThanTomorrow/calva/issues/1554)
+- [Apply basic typescript eslint rules](https://github.com/BetterThanTomorrow/calva/issues/1536)
+- Fix: [”Resolve macro as...” code action produces unreadable text in pop up](https://github.com/BetterThanTomorrow/calva/issues/1539)
+- Fix: [Format on save](https://github.com/BetterThanTomorrow/calva/issues/1556)
+
 ## [2.0.244] - 2022-02-20
 - [Add custom hover snippets](https://github.com/BetterThanTomorrow/calva/issues/1471)
 - [Add option to read cljfmt config from clojure-lsp](https://github.com/BetterThanTomorrow/calva/issues/1545)